+++ src/runtime_src/xdp/profile/database/static_info/xclbin_info.h
@@ -241,9 +241,9 @@ namespace xdp {
     std::vector<Monitor*> getUserAIMsWithTrace(XclbinInfo* xclbin) ;
     std::vector<Monitor*> getUserASMsWithTrace(XclbinInfo* xclbin) ;
 
-    void addTraceGMIO(uint32_t id, uint16_t col, uint16_t num,
-                                uint16_t stream, uint16_t len) ;
-    void addAIECounter(uint32_t i, uint16_t col, uint16_t r,
+    void addTraceGMIO(uint32_t id, uint8_t col, uint8_t num,
+                                uint8_t stream, uint8_t len) ;
+    void addAIECounter(uint32_t i, uint8_t col, uint8_t r,
                                  uint8_t num, uint16_t start, uint16_t end,
                                  uint8_t reset, uint32_t load, double freq,
                                  const std::string& mod,
